+++ tests/test_event_processor.py
@@ -103,7 +103,7 @@ EXPECTED_MEMBERS_ONLY_TYPE_FOR_PAID = {
 }
 
 EXPECTED_REPORT = """AC: Mending Monday (Test Event)
-    2020-11-16 19:00 to 2020-11-16 21:00
+    2020-11-16 19:00 to 21:00
     Meetup RSVP    $  0.00   3 registered on Meetup
     RSVP           $  0.00   unlimited