+++ tests/test_reporter.py
@@ -8,8 +8,13 @@ from datetime import datetime
 import io
 import pytest
 
-EXPECTED_FREE_EVENT_REPORT = (
-    "AC: Mending Monday (Test Event)\n" "    2020-11-09 19:00 to 2020-11-09 21:00\n"
+EXPECTED_FREE_EVENT_NAME = "AC: Mending Monday (Test Event)\n"
+EXPECTED_FREE_EVENT_DATES = "    2020-11-09 19:00 to 21:00\n"
+EXPECTED_LONGER_FREE_EVENT_DATES = "    2020-11-09 19:00 to 2020-11-10 01:00\n"
+
+EXPECTED_FREE_EVENT_REPORT = EXPECTED_FREE_EVENT_NAME + EXPECTED_FREE_EVENT_DATES
+EXPECTED_LONGER_FREE_EVENT_REPORT = (
+    EXPECTED_FREE_EVENT_NAME + EXPECTED_LONGER_FREE_EVENT_DATES
 )
 
 EXPECTED_MEETUP_RSVP_FREE = "    Meetup RSVP    $  0.00   0 registered on Meetup\n"
@@ -27,6 +32,14 @@ def free_apricot_event(free_meetup_event):
     """Return a free Wild Apricot event."""
     return MeetupToApricotEventAdaptor(
         free_meetup_event, EXPECTED_FREE_PHOTO_PATH, EXPECTED_FREE_TAGS
+    )
+
+
+@pytest.fixture()
+def longer_free_apricot_event(longer_free_meetup_event):
+    """Return a longer free Wild Apricot event."""
+    return MeetupToApricotEventAdaptor(
+        longer_free_meetup_event, EXPECTED_FREE_PHOTO_PATH, EXPECTED_FREE_TAGS
     )
 
 
@@ -53,6 +66,13 @@ def test_report_event(event_report, free_apricot_event, output):
     event_report.add_event(free_apricot_event)
     event_report.report_event(output)
     assert output.getvalue() == EXPECTED_FREE_EVENT_REPORT
+
+
+def test_report_event_longer(event_report, longer_free_apricot_event, output):
+    """Test reporting on an event."""
+    event_report.add_event(longer_free_apricot_event)
+    event_report.report_event(output)
+    assert output.getvalue() == EXPECTED_LONGER_FREE_EVENT_REPORT
 
 
 def test_report_photo_name(event_report, output):